% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcaogeobr.R
\name{funcaogeobr}
\alias{funcaogeobr}
\title{Bioma}
\usage{
funcaogeobr()
}
\description{
Funcao retorna bioma pampa
}
\examples{
\dontrun{
library(teste)
c<-funcaogeobr()

}
}
\seealso{
Other funcoes gerais: 
\code{\link{funcaoshape}()}
}
\concept{funcoes gerais}
