% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcaoshape.R
\name{funcaoshape}
\alias{funcaoshape}
\title{Mapa Bage}
\usage{
funcaoshape()
}
\description{
Funcao retorna o formato do municipio de Bage
}
\examples{
\dontrun{
library(teste)
a<-funcaobage()
}
}
\seealso{
Other funcoes gerais: 
\code{\link{funcaogeobr}()}
}
\concept{funcoes gerais}
